requirements.txt
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
